# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/courses": {
      "get": {
        "operationId": "CoursesController_findAll",
        "summary": "Daftar kursus (katalog publik)",
        "description": "Mendukung filter kategori, level, rentang harga, pencarian teks, dan sorting. Publik hanya melihat kursus berstatus PUBLISHED.",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Kata kunci pencarian",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filter slug kategori",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "level",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "BEGINNER",
                "INTERMEDIATE",
                "ADVANCED"
              ],
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "description": "Harga minimum (IDR)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "description": "Harga maksimum (IDR)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "free",
            "required": false,
            "in": "query",
            "description": "Hanya kursus gratis",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "featured",
            "required": false,
            "in": "query",
            "description": "Hanya kursus unggulan",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "popular",
              "enum": [
                "popular",
                "newest",
                "rating",
                "price_asc",
                "price_desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Admin/instructor saja — publik selalu PUBLISHED",
            "schema": {
              "enum": [
                "DRAFT",
                "PUBLISHED",
                "ARCHIVED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Courses"
        ]
      },
      "post": {
        "operationId": "CoursesController_create",
        "summary": "Buat kursus baru",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Courses"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/courses/{slug}": {
      "get": {
        "operationId": "CoursesController_findOne",
        "summary": "Detail kursus beserta kurikulum",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "example": "playwright-zero-to-expert",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Courses"
        ]
      }
    },
    "/api/v1/courses/{id}": {
      "patch": {
        "operationId": "CoursesController_update",
        "summary": "Perbarui kursus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Courses"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "CoursesController_remove",
        "summary": "Arsipkan kursus",
        "description": "Soft delete — status menjadi ARCHIVED, data pembelian tetap utuh.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Courses"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/courses/{id}/pricing": {
      "patch": {
        "operationId": "CoursesController_updatePricing",
        "summary": "Ubah harga satu kursus",
        "description": "Dipakai halaman CMS → Pricing.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCoursePricingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Courses"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/courses/pricing/bulk": {
      "patch": {
        "operationId": "CoursesController_bulkPricing",
        "summary": "Ubah harga banyak kursus sekaligus (transaksional)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkPricingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "updated": 12
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Courses"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/courses/{id}/refresh-stats": {
      "post": {
        "operationId": "CoursesController_refreshStats",
        "summary": "Hitung ulang rating / jumlah siswa / jumlah lesson",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Courses"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/categories": {
      "get": {
        "operationId": "CategoriesController_findAll",
        "summary": "Daftar kategori beserta jumlah kursus terbit",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Categories"
        ]
      },
      "post": {
        "operationId": "CategoriesController_create",
        "summary": "Tambah kategori",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/categories/{id}": {
      "patch": {
        "operationId": "CategoriesController_update",
        "summary": "Perbarui kategori",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "CategoriesController_remove",
        "summary": "Hapus kategori (gagal bila masih dipakai kursus)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/content/pages/{slug}": {
      "get": {
        "operationId": "PublicContentController_page",
        "summary": "Halaman terbit beserta blok kontennya",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Content (Public)"
        ]
      }
    },
...
    "/api/v1/content/testimonials": {
      "get": {
        "operationId": "PublicContentController_testimonials",
        "summary": "Testimoni (opsional hanya yang unggulan)",
        "parameters": [
          {
            "name": "featured",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Content (Public)"
        ]
      }
    },
...
    "/api/v1/content/navigation": {
      "get": {
        "operationId": "PublicContentController_navigation",
        "summary": "Item menu yang terlihat",
        "parameters": [
          {
            "name": "menu",
            "required": false,
            "in": "query",
            "schema": {
              "example": "header",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Content (Public)"
        ]
      }
    },
...
    "/api/v1/leads": {
      "post": {
        "operationId": "LeadsController_create",
        "summary": "Kirim data calon peserta dari form landing page",
        "description": "Minimal salah satu dari email atau nomor telepon wajib diisi.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLeadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Leads"
        ]
      },
      "get": {
        "operationId": "LeadsController_list",
        "summary": "Daftar leads (admin)",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Kata kunci pencarian",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Leads"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/leads/{id}": {
      "patch": {
        "operationId": "LeadsController_update",
        "summary": "Tandai lead sudah ditindaklanjuti",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLeadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Leads"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "TestCraft Indonesia LMS API",
    "description": "REST API untuk platform LMS TestCraft Indonesia.\n\n### Autentikasi\nSemua endpoint memerlukan header `Authorization: Bearer <accessToken>`\nkecuali yang bertanda **Public**. Access token berlaku 15 menit;\ngunakan `POST /auth/refresh` untuk memperbarui (refresh token dirotasi\nsetiap kali dipakai).\n\n### Format error\n```json\n{\n  \"statusCode\": 400,\n  \"message\": [\"email harus berupa alamat email yang valid\"],\n  \"error\": \"Bad Request\",\n  \"path\": \"/api/v1/auth/register\",\n  \"timestamp\": \"2026-07-27T03:21:00.000Z\"\n}\n```\n\n### Paginasi\nEndpoint list menerima `?page=1&limit=20&q=kata-kunci` dan mengembalikan\n`{ data: [...], meta: { page, limit, total, totalPages, hasNext, hasPrev } }`.\n\n### Rate limit\n120 request per menit per IP.",
    "version": "1.0.0",
    "contact": {
      "name": "TestCraft Indonesia",
      "url": "https://wa.me/6282395568743",
      "email": "testcraftindonesia@gmail.com"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Local development"
    },
    {
      "url": "https://api.testcraft.id/api/v1",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCourseDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "example": "playwright-zero-to-expert"
          },
          "title": {
            "type": "string",
            "example": "Playwright Automation Testing from Zero to Expert"
          },
          "subtitle": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "level": {
            "type": "string",
            "enum": [
              "BEGINNER",
              "INTERMEDIATE",
              "ADVANCED"
            ]
          },
          "categoryId": {
            "type": "string",
            "description": "ID kategori"
          },
          "instructorId": {
            "type": "string",
            "description": "ID profil instruktur"
          },
          "priceIDR": {
            "type": "number",
            "example": 1250000
          },
          "priceUSD": {
            "type": "number",
            "example": 79
          },
          "compareAtIDR": {
            "type": "number",
            "example": 2500000,
            "description": "Harga coret"
          },
          "compareAtUSD": {
            "type": "number"
          },
          "isFree": {
            "type": "boolean",
            "default": false
          },
          "icon": {
            "type": "string",
            "example": "🎭"
          },
          "thumbnailUrl": {
            "type": "string"
          },
          "previewVideoUrl": {
            "type": "string"
          },
          "durationMin": {
            "type": "number",
            "example": 1920,
            "description": "Durasi total (menit)"
          },
          "outcomes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "prerequisites": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "targetAudience": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "ARCHIVED"
            ],
            "default": "DRAFT"
          },
          "isBestseller": {
            "type": "boolean"
          },
          "isFeatured": {
            "type": "boolean"
          },
          "seoTitle": {
            "type": "string"
          },
          "seoDescription": {
            "type": "string"
          }
        },
        "required": [
          "slug",
          "title",
          "description",
          "level",
          "categoryId",
          "instructorId"
        ]
      },
      "UpdateCourseDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "example": "playwright-zero-to-expert"
          },
          "title": {
            "type": "string",
            "example": "Playwright Automation Testing from Zero to Expert"
          },
          "subtitle": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "level": {
            "type": "string",
            "enum": [
              "BEGINNER",
              "INTERMEDIATE",
              "ADVANCED"
            ]
          },
          "categoryId": {
            "type": "string",
            "description": "ID kategori"
          },
          "instructorId": {
            "type": "string",
            "description": "ID profil instruktur"
          },
          "priceIDR": {
            "type": "number",
            "example": 1250000
          },
          "priceUSD": {
            "type": "number",
            "example": 79
          },
          "compareAtIDR": {
            "type": "number",
            "example": 2500000,
            "description": "Harga coret"
          },
          "compareAtUSD": {
            "type": "number"
          },
          "isFree": {
            "type": "boolean",
            "default": false
          },
          "icon": {
            "type": "string",
            "example": "🎭"
          },
          "thumbnailUrl": {
            "type": "string"
          },
          "previewVideoUrl": {
            "type": "string"
          },
          "durationMin": {
            "type": "number",
            "example": 1920,
            "description": "Durasi total (menit)"
          },
          "outcomes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "prerequisites": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "targetAudience": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "ARCHIVED"
            ],
            "default": "DRAFT"
          },
          "isBestseller": {
            "type": "boolean"
          },
          "isFeatured": {
            "type": "boolean"
          },
          "seoTitle": {
            "type": "string"
          },
          "seoDescription": {
            "type": "string"
          }
        }
      },
      "UpdateCoursePricingDto": {
        "type": "object",
        "properties": {
          "priceIDR": {
            "type": "number",
            "example": 1350000
          },
          "priceUSD": {
            "type": "number",
            "example": 85
          },
          "compareAtIDR": {
            "type": "number",
            "example": 2700000
          },
          "compareAtUSD": {
            "type": "number"
          },
          "isFree": {
            "type": "boolean"
          }
        },
        "required": [
          "priceIDR",
          "priceUSD"
        ]
      },
      "BulkPricingItemDto": {
        "type": "object",
        "properties": {
          "priceIDR": {
            "type": "number",
            "example": 1350000
          },
          "priceUSD": {
            "type": "number",
            "example": 85
          },
          "compareAtIDR": {
            "type": "number",
            "example": 2700000
          },
          "compareAtUSD": {
            "type": "number"
          },
          "isFree": {
            "type": "boolean"
          },
          "courseId": {
            "type": "string"
          }
        },
        "required": [
          "priceIDR",
          "priceUSD",
          "courseId"
        ]
      },
      "BulkPricingDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BulkPricingItemDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Automation Testing"
          },
          "slug": {
            "type": "string",
            "example": "automation-testing"
          },
          "icon": {
            "type": "string",
            "example": "🤖"
          },
          "order": {
            "type": "number",
            "default": 0
          },
          "featured": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Automation Testing"
          },
          "slug": {
            "type": "string",
            "example": "automation-testing"
          },
          "icon": {
            "type": "string",
            "example": "🤖"
          },
          "order": {
            "type": "number",
            "default": 0
          },
          "featured": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "CreateLeadDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Rizky Ananda"
          },
          "email": {
            "type": "string",
            "example": "rizky@example.com"
          },
          "phone": {
            "type": "string",
            "example": "+6281234567890"
          },
          "interest": {
            "type": "string",
            "example": "Playwright Automation"
          },
          "message": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "default": "landing"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateLeadDto": {
        "type": "object",
        "properties": {
          "handled": {
            "type": "boolean"
          }
        },
        "required": [
          "handled"
        ]
      }
    }
  }
}